Split subrouter CLI implementation
